+++ src/urllib3/util/_async/ssl_.py
@@ -122,9 +122,11 @@ async def ssl_wrap_socket(
                 except OSError as e:
                     raise SSLError(e) from e
 
-            elif ssl_context is None and hasattr(context, "load_default_certs"):
+            elif hasattr(context, "load_default_certs"):
+                store_stats = context.cert_store_stats()
                 # try to load OS default certs; works well on Windows.
-                context.load_default_certs()
+                if "x509_ca" not in store_stats or not store_stats["x509_ca"]:
+                    context.load_default_certs()
 
             # Attempt to detect if we get the goofy behavior of the
             # keyfile being encrypted and OpenSSL asking for the